casdcxcz

+++ SafraVisionAPI/Models/ClienteModel.cs
@@ -5,7 +5,7 @@
         public int idCliente { get; set; }
         public String? nomeCliente { get; set; }
         public String? descricao { get; set; }
-        public int numeroTelefone { get; set; }
+        public Int64 numeroTelefone { get; set; }
 
     }
 }
